pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pck"
version = "0.0.0"
description = "The Universal Language Runner"
authors = [{name = "JusteLeo"}]
requires-python = ">=3.8"
dependencies = [
    "typer",
    "rich",
    "requests",
    "pyyaml"
]

[project.scripts]
pck = "main:app"

[tool.setuptools]
py-modules = ["main"]
packages = ["src"]